Added new reducer for handling lastIndexed
