+++ amundsen_application/static/js/ducks/lastIndexed/api/v0.ts
@@ -1,0 +1,15 @@
+import axios, { AxiosResponse, AxiosError } from 'axios';
+
+export const API_PATH = '/api/metadata/v0';
+
+type MessageAPI = { msg: string };
+
+export type LastIndexedAPI = { timestamp: string } & MessageAPI;
+
+export function getLastIndexed() {
+  return axios
+    .get(`${API_PATH}/get_last_indexed`)
+    .then((response: AxiosResponse<LastIndexedAPI>) => {
+      return response.data.timestamp;
+    });
+}
