+++ amundsen_application/static/js/ducks/tableMetadata/api/v0.ts
@@ -31,7 +31,7 @@ export type TableData = TableMetadata & {
   tags: Tag[];
 };
 export type DescriptionAPI = { description: string } & MessageAPI;
-export type LastIndexedAPI = { timestamp: string } & MessageAPI;
+// export type LastIndexedAPI = { timestamp: string } & MessageAPI;
 export type PreviewDataAPI = { previewData: PreviewData } & MessageAPI;
 export type TableDataAPI = { tableData: TableData } & MessageAPI;
 export type RelatedDashboardDataAPI = {
@@ -173,13 +173,13 @@ export function updateColumnDescription(
   });
 }
 
-export function getLastIndexed() {
-  return axios
-    .get(`${API_PATH}/get_last_indexed`)
-    .then((response: AxiosResponse<LastIndexedAPI>) => {
-      return response.data.timestamp;
-    });
-}
+// export function getLastIndexed() {
+//   return axios
+//     .get(`${API_PATH}/get_last_indexed`)
+//     .then((response: AxiosResponse<LastIndexedAPI>) => {
+//       return response.data.timestamp;
+//     });
+// }
 
 export function getPreviewData(queryParams: PreviewQueryParams) {
   return axios({
