+++ amundsen_application/static/js/ducks/tableMetadata/reducer.ts
@@ -24,9 +24,9 @@ import {
   GetColumnDescriptionRequest,
   UpdateColumnDescription,
   UpdateColumnDescriptionRequest,
-  GetLastIndexed,
-  GetLastIndexedRequest,
-  GetLastIndexedResponse,
+  // GetLastIndexed,
+  // GetLastIndexedRequest,
+  // GetLastIndexedResponse,
   GetPreviewData,
   GetPreviewDataRequest,
   GetPreviewDataResponse,
@@ -66,7 +66,7 @@ export const initialTableDataState: TableMetadata = {
 
 export const initialState: TableMetadataReducerState = {
   isLoading: true,
-  lastIndexed: null,
+  // lastIndexed: null,
   preview: initialPreviewState,
   statusCode: null,
   tableData: initialTableDataState,
@@ -231,22 +231,22 @@ export function updateColumnDescription(
   };
 }
 
-export function getLastIndexed(): GetLastIndexedRequest {
-  return { type: GetLastIndexed.REQUEST };
-}
-export function getLastIndexedFailure(): GetLastIndexedResponse {
-  return { type: GetLastIndexed.FAILURE };
-}
-export function getLastIndexedSuccess(
-  lastIndexedEpoch: number
-): GetLastIndexedResponse {
-  return {
-    type: GetLastIndexed.SUCCESS,
-    payload: {
-      lastIndexedEpoch,
-    },
-  };
-}
+// export function getLastIndexed(): GetLastIndexedRequest {
+//   return { type: GetLastIndexed.REQUEST };
+// }
+// export function getLastIndexedFailure(): GetLastIndexedResponse {
+//   return { type: GetLastIndexed.FAILURE };
+// }
+// export function getLastIndexedSuccess(
+//   lastIndexedEpoch: number
+// ): GetLastIndexedResponse {
+//   return {
+//     type: GetLastIndexed.SUCCESS,
+//     payload: {
+//       lastIndexedEpoch,
+//     },
+//   };
+// }
 
 export function getPreviewData(
   queryParams: PreviewQueryParams
@@ -286,7 +286,7 @@ export interface TableMetadataReducerState {
     errorMessage?: string;
   };
   isLoading: boolean;
-  lastIndexed: number;
+  // lastIndexed: number;
   preview: {
     data: PreviewData;
     status: number | null;
@@ -311,10 +311,11 @@ export default function reducer(
         },
       };
     case GetTableData.REQUEST:
-      return {
-        ...initialState,
-        lastIndexed: state.lastIndexed,
-      };
+      return initialState;
+    // return {
+    //   ...initialState,
+    //   lastIndexed: state.lastIndexed,
+    // };
     case GetTableData.FAILURE:
       return {
         ...state,
@@ -344,13 +345,13 @@ export default function reducer(
         ...state,
         tableData: (<GetColumnDescriptionResponse>action).payload.tableMetadata,
       };
-    case GetLastIndexed.FAILURE:
-      return { ...state, lastIndexed: null };
-    case GetLastIndexed.SUCCESS:
-      return {
-        ...state,
-        lastIndexed: (<GetLastIndexedResponse>action).payload.lastIndexedEpoch,
-      };
+    // case GetLastIndexed.FAILURE:
+    //   return { ...state, lastIndexed: null };
+    // case GetLastIndexed.SUCCESS:
+    //   return {
+    //     ...state,
+    //     lastIndexed: (<GetLastIndexedResponse>action).payload.lastIndexedEpoch,
+    //   };
     case GetPreviewData.FAILURE:
     case GetPreviewData.SUCCESS:
       return { ...state, preview: (<GetPreviewDataResponse>action).payload };
