+++ amundsen_application/static/js/ducks/tableMetadata/types.ts
@@ -116,20 +116,20 @@ export interface UpdateColumnDescriptionResponse {
   type: UpdateColumnDescription.SUCCESS | UpdateColumnDescription.FAILURE;
 }
 
-export enum GetLastIndexed {
-  REQUEST = 'amundsen/tableMetadata/GET_LAST_UPDATED_REQUEST',
-  SUCCESS = 'amundsen/tableMetadata/GET_LAST_UPDATED_SUCCESS',
-  FAILURE = 'amundsen/tableMetadata/GET_LAST_UPDATED_FAILURE',
-}
-export interface GetLastIndexedRequest {
-  type: GetLastIndexed.REQUEST;
-}
-export interface GetLastIndexedResponse {
-  type: GetLastIndexed.SUCCESS | GetLastIndexed.FAILURE;
-  payload?: {
-    lastIndexedEpoch: number;
-  };
-}
+// export enum GetLastIndexed {
+//   REQUEST = 'amundsen/tableMetadata/GET_LAST_UPDATED_REQUEST',
+//   SUCCESS = 'amundsen/tableMetadata/GET_LAST_UPDATED_SUCCESS',
+//   FAILURE = 'amundsen/tableMetadata/GET_LAST_UPDATED_FAILURE',
+// }
+// export interface GetLastIndexedRequest {
+//   type: GetLastIndexed.REQUEST;
+// }
+// export interface GetLastIndexedResponse {
+//   type: GetLastIndexed.SUCCESS | GetLastIndexed.FAILURE;
+//   payload?: {
+//     lastIndexedEpoch: number;
+//   };
+// }
 
 export enum GetPreviewData {
   REQUEST = 'amundsen/preview/GET_PREVIEW_DATA_REQUEST',
